Uart: implement non-blocked writing for all drivers
In non-blocking mode, the write function now returns immediately if
'tx_avail()' reports that writing of a single character is not possible.

As 'out_char()' is not accessible via the Uart_base class, the behavior
of that function changes to no longer wait until 'tx_avail()' returns
true. This is now done in 'write()'. While at it, unify the start value
of the poll timeout counter between all drivers.

Note that non-blocking mode might need more adaptions in certain drivers
because the `tx_avail()` function of these drivers is sometimes
implemented too conservative.

Change-Id: I733c537bc9a1b9471c70603802183e53063f838f