Add inline asm mmio access functions
Use inline assembly to perform MMIO register accesses on x86, amd64, arm
and arm64. The idea is to have the choice of access instructions under
our control.  This is important especially for platforms that can run as
guest under KVM, which uses incomplete instruction decoder and thus can
handle only certain instructions.

The generic implementation is retained for mips and other platforms.

Change-Id: I28529e9c8681e6d260d4dc41d186ced460461b5e