x86: fix register constrains for 8-bit read/write
When reading or writing a single byte from or to memory, the involved
register needs to be 8-bit as well. Specifying 'r' would allow to use
any GP register. This includes ebp, esi and edi which are can be only
used with a size of 16-bit or 32-bit.

This fixes a weird error message when compiling certain code for x86:

  error: unsupported size for integer register

Change-Id: Ifac9473e4d1fa302bb595a5219cea7ba7bcde008