uart: force strict alignment
The uart drivers used for code where the MMU could be disabled, for
example in bootstrap. In that case, the generated code has to follow
natural memory alignment rules, for example a 16-byte access needs to be
16-byte aligned architectures.

The proper compile switch for this is -mstrict-align. This option is
only supported on ARM64.

Change-Id: I2c8a817a2f3bded9f450176e799c9c7fabd0d00a