hpet: fix Clang warning
Clang emits the following warning:

  hpet:41:71: warning:
    taking address of packed member '_int_status' of class or structure
    'L4drivers::Hpet' may result in an unaligned pointer value
    [-Waddress-of-packed-member]
    __typeof(_int_status) dummy
         = *(volatile __typeof(_int_status) *)&_int_status;
                                              ^~~~~~~~~~~

This warning is overcautious because `_int_status` is actually not
mis-aligned. However, the `Hpet` structure doesn't need to be "packed"
because all members are naturally aligned.

Therefore, remove the "packed" attribute and add a static assertion
asserting on the size of the structure. This fixes the Clang warning
using a conservative approach.

The `_pad4` member is not necessary but pad the structure to the next
relevant member ("timer 0 configuration and capabilities register").

Change-Id: I94ebc4edd0d4795f69318f9b8bd04d530899efb5