nand: Replace usage of loff_t with u32
The choice of loff_t as a type for an address is kind of peculiar, since
loff_t is a type "used for describing file sizes" that was introduced by
the Linux kernel.

Since all addresses used in the nand code seem to be 32-bit, and in
other parts of the code u32 is used to store addresses, we replace
loff_t with u32.

Change-Id: I8b909c387bbec3e9fd236b00d95fc206cd2e9499
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>