Remove useless L4_NO_RTTI define
When compiling without RTTI (i.e. -fno-rtti) we do not need L4_NO_RTTI.
We only need it when we compile with RTTI in principle, but do not want
to include the L4 RTTI things. So remove L4_NO_RTTI in all cases where
we also specify -fno-rtti when compiling.

See l4sys/include/__typeinfo.h for where this is used.

Change-Id: I3358e8840aa5e49bc14af7c8fc19a59d28b5653f
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>